+++ spec/selinux_module_spec.rb
@@ -45,7 +45,8 @@ alsa       1.12.2
   end
 
   it 'installs the dependency packages' do
-    expect(chef_run).to(install_package('make, policycoreutils, selinux-policy-devel'))
+    pkgs = os.debian? ? %w(make policycoreutils selinux-policy-dev) : %w(make policycoreutils selinux-policy-devel)
+    expect(chef_run).to(install_package(pkgs))
   end
 
   it 'logs the steps taken' do