+++ CHANGELOG.md
@@ -4,9 +4,15 @@ This file is used to list changes made in each version of the selinux cookbook.
 
 ## Unreleased
 
-- Refactor as a library cookbook
-- Restore support for Debian/Ubuntu
-- Expand test platforms
+### Note: With version 5.0.0 the default recipe has been removed
+
+- Major refactoring
+- Restore support for Debian based distros
+- All resources now use unified_mode
+- Added selinux_boolean resource
+- Remove attributes and default recipe
+  - Replaced with a set of bare recipes for the three selinux states
+- Add automatic restart function to `selinux_state` resource
 
 ## 4.0.0 - *2021-07-21*